requirement.txt
streamlit==1.50.0
tensorflow-cpu==2.15.0
pillow==11.3.0
numpy==2.3.3
